feat: make all adapters available on all platforms
