+++ src/platform.node.ts
@@ -21,6 +21,9 @@ export function baseFetchConfig(apiRoot: string) {
     } else return {};
 }
 
+// === Default webhook adapter
+export const defaultAdapter = "express";
+
 // === InputFile handling and File augmenting
 // Accessor for file data in `InputFile` instances
 export const toRaw = Symbol("InputFile data");
