+++ src/platform.web.ts
@@ -8,6 +8,8 @@ export { readableStreamFromIterable as itrToStream } from "https://deno.land/std
 // === Base configuration for `fetch` calls
 export const baseFetchConfig = (_apiRoot: string) => ({});
 
+export const defaultAdapter = "cloudflare";
+
 // === InputFile handling and File augmenting
 // Accessor for file data in `InputFile` instances
 export const toRaw = Symbol("InputFile data");
